Base V1.8
Preliminary implementation of base models into category building
constants for category is working
test.py has sample code testing the building of the model by iterating through categories and this is working.
Currently the building of the family object is failing in the api_families.py, so the model is not updating as desired.